bugfix, inventoryUpdate
-Fixed a bug where player would get twice as  much health from healths and potions as it should
(removed healing methods from OnTriggerStay)
-Added some basic Inventory UI to player screen
-retyped float variables to ints.
-created health potions, (origin: unity asset store) that can be dropped by enemies upon dying. it restores 25 health.
-Added item.cs and inventory.cs with no implemented real function yet.
-reworked  health adding system for readablitiy. (dynamic type that gets decided in runtime wether it is Heal or HealthPotion)

+++ Assets/scripts/Heal.cs
@@ -9,8 +9,8 @@ public class Heal : MonoBehaviour
     float timeWhenDestroy = 15f;
     Vector3 rotation = new Vector3(0, 45, 0);
 
-    private float healAmount = 70f;
-    public float HealAmount
+    private int healAmount = 70;
+    public int HealAmount
     {
         get
         {
